[compiler] Add wrapper functions to wrap change-detection storage and loading from the memo cache
Summary:
We may wish to perform some additional computation on values when they enter or exit the memo cache in change detection mode (e.g. make a deep copy, restore the original value). This builds support for doing so.

In addition, it drops the "ForDebugging" part of the flag name and makes it compatible with "disableMemoization".

ghstack-source-id: cda9a0afb426dadb48c40d23c53ce0f00f960d02
Pull Request resolved: https://github.com/facebook/react/pull/30181